Add flexible and strict type mapping BNFC --> Java, untested
Issue 153. When using the Integer or Double built-in token classes
if e.g. an Integer whose binary representation exceeds 32 bits
in length, a java.lang.NumberFormatException is thrown.

To fix this the new type TypeMapping is created, being essentially
a mapping between BNFC token categories to Java corresponding types.

This type is added to the signature of all functions that need to
access such mapping.