Replace type Either a b (and related) to RhsRuleElement a b whenever the former is used to model an element of the right hand side of a rule; enable Antlr Java backend to generate indentation aware parsers.
The type Either a b was used in the modified files to represent, essentially,
b=terminals corresponding to anonymous tokens in the .cf file,
a=all other nonterminals.
To enable indentation-aware ANTLR grammars, I had to artificially inject
three kinds of tokens, which are unreachable from .cf files.
For this reason I needed a device to differentiate one more class of tokens
which are indentation tokens (see RhsRuleElement definition in CF.hs).
This led to a cascade of changes in other files which are anyway unaware of
indentation issues (or they solve them otherwise using the Layout pragmas
like in Haskell)
Nota bene: To implement the indentation-aware parser generation schema
I implemented for ANTLR it is necessary that your parser generator
supports the presence of semantic actions before and after any non-terminal
representing a list of something.